Revert "Merge branch 'dev_entitycoreframework' into devfeature"
This reverts commit 9779882756aeec9620452a0718effef404ab6f57, reversing
changes made to cd228a9344a81081c749de5646191b0d8f9005a8.

+++ StreamerBotLib/Overlay/Models/OverlayActionType.cs
@@ -37,7 +37,7 @@ namespace StreamerBotLib.Overlay.Models
         /// <summary>
         /// The user provided length of the notification.
         /// </summary>
-        public int Duration { get; set; } = 0;
+        public decimal Duration { get; set; } = 0;
 
         /// <summary>
         /// The path to an image used in the event.
@@ -88,7 +88,7 @@ namespace StreamerBotLib.Overlay.Models
                 if (strings.Count > 6)
                 {
                     OverlayTypes type = (OverlayTypes)Enum.Parse(typeof(OverlayTypes), strings.Dequeue());
-                    int Duration = int.Parse(strings.Dequeue());
+                    decimal Duration = decimal.Parse(strings.Dequeue());
                     string User = strings.Dequeue();
                     string Msg = strings.Dequeue();
                     string action = strings.Dequeue();
